// src/users/constants/scope.enum.ts
// src/users/constants/scope.enum.ts

export enum Scope {
  ADMIN_PORTAL = 'admin_portal',
  MERCHANT_WEB = 'merchant_web',
  MERCHANT_ANDROID = 'merchant_android',
  MERCHANT_IOS = 'merchant_ios',
  MERCHANT_CLOVER = 'merchant_clover',
  CLOVER_SDK = 'clover_sdk',
  PUBLIC_CHECKOUT = 'public_checkout',
  EXTERNAL_API = 'external_api',
}